style(CanvasSettings): change inline classNames for tw styled component
